second refactoring iteration
