+++ src/binder/Cargo.toml
@@ -3,7 +3,14 @@ name = "binder"
 version = "0.1.0"
 authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
 edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+publish = false
 
 [dependencies]
+protocol = { path = "../protocol" }
+
+bigdecimal = "0.1.2"
+log = "0.4.11"
+sqlparser = { version = "0.6.1", features = ["bigdecimal"] }
+
+[dev-dependencies]
+rstest = "0.6.4"
