+++ src/node/Cargo.toml
@@ -13,11 +13,13 @@ name = "database"
 path = "src/bin.rs"
 
 [dependencies]
+binder = { path = "../binder" }
 data_manager = { path = "../data_manager" }
 kernel = { path = "../kernel" }
 parser = { path = "../parser" }
 protocol = { path = "../protocol" }
-sql_engine = { path = "../sql_engine" }
+query_executor = { path = "../query_executor" }
+query_planner = { path = "../query_planner" }
 sql_model = { path = "../sql_model" }
 
 async-dup = "1.2.1"
@@ -26,4 +28,8 @@ blocking = "1.0.0"
 futures-lite = "1.7.0"
 log = "0.4.8"
 simple_logger = { version = "1.9.0", default-features = false }
+sqlparser = { version = "0.6.1", features = ["bigdecimal"] }
 smol = "1.0.0"
+
+[dev-dependencies]
+rstest = "0.6.4"
