+++ src/node/src/node.rs
@@ -25,12 +25,9 @@ use std::{
 use async_dup::Arc as AsyncArc;
 use async_io::Async;
 
+use crate::query_engine::QueryEngine;
 use data_manager::DataManager;
-use parser::QueryParser;
-use protocol::results::QueryError;
-use protocol::session::Session;
-use protocol::{Command, ProtocolConfiguration, Receiver, Sender};
-use sql_engine::QueryExecutor;
+use protocol::{ProtocolConfiguration, Receiver, Sender};
 
 const PORT: u16 = 5432;
 const HOST: [u8; 4] = [0, 0, 0, 0];
@@ -64,13 +61,9 @@ pub fn start() {
                     let state = state.clone();
                     let storage = storage.clone();
                     let sender = Arc::new(sender);
-                    let mut query_executor = QueryExecutor::new(storage.clone(), sender.clone());
-                    let mut query_parser = QueryParser::new(sender.clone(), storage.clone());
+                    let mut query_engine = QueryEngine::new(sender.clone(), storage.clone());
                     log::debug!("ready to handle query");
-
-                    let sender_clone = sender.clone();
                     smol::spawn(async move {
-                        let mut session = Session::default();
                         loop {
                             match receiver.receive().await {
                                 Err(e) => {
@@ -83,89 +76,14 @@ pub fn start() {
                                     state.store(STOPPED, Ordering::SeqCst);
                                     return;
                                 }
-                                Ok(Ok(Command::Bind {
-                                    portal_name,
-                                    statement_name,
-                                    param_formats,
-                                    raw_params,
-                                    result_formats,
-                                })) => {
-                                    match session.get_prepared_statement(&statement_name) {
-                                        Some(prepared_statement) => {
-                                            let param_types = prepared_statement.param_types();
-                                            if param_types.len() != raw_params.len() {
-                                                let message = format!(
-                                                    "Bind message supplies {actual} parameters, but prepared statement \"{name}\" requires {expected}",
-                                                    name = statement_name,
-                                                    actual = raw_params.len(),
-                                                    expected = param_types.len()
-                                                );
-                                                sender_clone.send(Err(QueryError::protocol_violation(message))).expect("To Send Error to Client");
-                                            }
-                                            match query_executor.bind_prepared_statement_to_portal(
-                                                &prepared_statement,
-                                                param_formats.as_ref(),
-                                                raw_params.as_ref(),
-                                                result_formats.as_ref(),
-                                            ) {
-                                                Ok((new_stmt, result_formats)) => {
-                                                    session.set_portal(
-                                                        portal_name.to_owned(),
-                                                        statement_name.to_owned(),
-                                                        new_stmt,
-                                                        result_formats,
-                                                    );
-                                                }
-                                                Err(error) => log::error!("{:?}", error),
-                                            }
-                                        },
-                                        None => {
-                                            sender_clone
-                                                .send(Err(QueryError::prepared_statement_does_not_exist(
-                                                    statement_name,
-                                                )))
-                                                .expect("To Send Error to Client");
-                                        }
+                                Ok(Ok(command)) => match query_engine.execute(command) {
+                                    Ok(()) => {
+                                        receiver.ready_for_query().await.expect("Ok");
                                     }
-                                }
-                                Ok(Ok(Command::Continue)) => {
-                                    receiver.ready_for_query().await.expect("Ok");
-                                }
-                                Ok(Ok(Command::DescribeStatement { name })) => {
-                                    match query_executor.describe_prepared_statement(name.as_str()) {
-                                        Ok(()) => {}
-                                        Err(error) => log::error!("{:?}", error),
+                                    Err(()) => {
+                                        break;
                                     }
-                                }
-                                Ok(Ok(Command::Execute { portal_name, max_rows })) => {
-                                    match query_executor.execute_portal(portal_name.as_str(), max_rows) {
-                                        Ok(()) => {}
-                                        Err(error) => log::error!("{:?}", error),
-                                    }
-                                }
-                                Ok(Ok(Command::Flush)) => query_executor.flush(),
-                                Ok(Ok(Command::Parse {
-                                    statement_name,
-                                    sql,
-                                    param_types,
-                                })) => {
-                                    match query_parser.parse_prepared_statement(sql.as_str(), param_types.as_ref()) {
-                                        Ok(statement) => {
-                                            session.set_prepared_statement(statement_name, statement);
-                                        }
-                                        Err(error) => log::error!("{:?}", error),
-                                    }
-                                }
-                                Ok(Ok(Command::Query { sql })) => {
-                                    if let Ok(statement) = query_parser.parse(sql.as_str()) {
-                                        query_executor.execute(&statement);
-                                        query_executor.flush();
-                                    }
-                                }
-                                Ok(Ok(Command::Terminate)) => {
-                                    log::debug!("closing connection with client");
-                                    break;
-                                }
+                                },
                             }
                         }
                     })
