+++ src/node/src/query_engine/tests/delete.rs
