+++ src/query_executor/src/ddl/drop_schema.rs
