+++ src/query_executor/src/ddl/drop_table.rs
